// packages/protocol/typescript/generated/yiru/runtime/v1/git_generation_pb.d.ts
export declare type GitPullRequestFields = Message<"yiru.runtime.v1.GitPullRequestFields"> & {
  /**
   * @generated from field: string base = 1;
   */
  base: string;

  /**
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * @generated from field: string body = 3;
   */
  body: string;

  /**
   * @generated from field: bool draft = 4;
   */
  draft: boolean;
};